Backward covariate search

Backward covariate search

Usage

backwardSearch(
  varsVec,
  covarsVec,
  catvarsVec = NULL,
  fitorig,
  fitupdated,
  pVal = 0.01,
  reFitCovars = FALSE,
  outputDir,
  restart = FALSE
)

Arguments

varsVec

character vector of variables that need to be added

covarsVec

character vector of covariates that need to be added

catvarsVec

character vector of categorical covariates that need to be added

fitorig

the original 'fit' object before forward search

fitupdated

the updated 'fit' object, if any, after the forward search

pVal

p-value that should be used for selecting covariates in the forward search

reFitCovars

if the covariates should be added before performing backward search - useful for directly performing backward search without forward search; default is FALSE

outputDir

the name of the output directory that stores the covariate search result

restart

a boolean that controls if the search should be restarted; default is FALSE

Value

returns the updated 'fit' object at the end of the backward search and a table of information for all the covariates tested

Author

Vipul Mann, Matthew Fidler, Vishal Sarsani