nlmixr2extra: Nonlinear Mixed Effects Models in Population PK/PD, Extra Support Functions

Fit and compare nonlinear mixed-effects models in differential equations with flexible dosing information commonly seen in pharmacokinetics and pharmacodynamics (Almquist, Leander, and Jirstrand 2015 doi:10.1007/s10928-015-9409-1 ). Differential equation solving is by compiled C code provided in the 'rxode2' package (Wang, Hallow, and James 2015 doi:10.1002/psp4.12052 ). This package is for support functions like preconditioned fits doi:10.1208/s12248-016-9866-5 , boostrap and stepwise covariate selection.

See also

Author

Maintainer: Matthew Fidler matthew.fidler@gmail.com (ORCID)

Authors:

Matthew Fidler matthew.fidler@gmail.com (ORCID)

Vipul Mann vm2583@columbia.edu

Vishal Sarsani vsarsani@umass.edu (ORCID)

Bill Denney wdenney@humanpredictions.com (ORCID)

Other contributors:

Christian Bartels christian.bartels@novartis.com [contributor]

Omar Elashkar omar.i.elashkar@gmail.com (ORCID) [contributor]