Convert nlmixr compatible data to monolix compatible data (if possible)

Convert nlmixr compatible data to monolix compatible data (if possible)

Usage

nlmixrDataToMonolix(model, data, table = nlmixr2est::tableControl())

Arguments

model

rxode2 model for conversion

data

Input dataset

table

is the table control; this is mostly to figure out if there are additional columns to keep.

Value

Monolix compatible dataset

Author

Matthew L. Fidler

Examples

pk.turnover.emax3 <- function() {
  ini({
    tktr <- log(1)
    tka <- log(1)
    tcl <- log(0.1)
    tv <- log(10)
    ##
    eta.ktr ~ 1
    eta.ka ~ 1
    eta.cl ~ 2
    eta.v ~ 1
    prop.err <- 0.1
    pkadd.err <- 0.1
    ##
    temax <- logit(0.8)
    tec50 <- log(0.5)
    tkout <- log(0.05)
    te0 <- log(100)
    ##
    eta.emax ~ .5
    eta.ec50  ~ .5
    eta.kout ~ .5
    eta.e0 ~ .5
    ##
    pdadd.err <- 10
  })
  model({
    ktr <- exp(tktr + eta.ktr)
    ka <- exp(tka + eta.ka)
    cl <- exp(tcl + eta.cl)
    v <- exp(tv + eta.v)
    emax = expit(temax+eta.emax)
    ec50 =  exp(tec50 + eta.ec50)
    kout = exp(tkout + eta.kout)
    e0 = exp(te0 + eta.e0)
    ##
    DCP = center/v
    PD=1-emax*DCP/(ec50+DCP)
    ##
    effect(0) = e0
    kin = e0*kout
    ##
    d/dt(depot) = -ktr * depot
    d/dt(gut) =  ktr * depot -ka * gut
    d/dt(center) =  ka * gut - cl / v * center
    d/dt(effect) = kin*PD -kout*effect
    ##
    cp = center / v
    cp ~ prop(prop.err) + add(pkadd.err)
    effect ~ add(pdadd.err) | pca
  })
}

nlmixrDataToMonolix(pk.turnover.emax3, nlmixr2data::warfarin)
#>  
#>     ID  TIME EVID   AMT II    DV ADM YTYPE SS nlmixrRowNums
#> 1    1   0.0    1 100.0  0    NA   1     0  0             1
#> 2    1   0.5    0    NA  0   0.0   5     1  0             2
#> 3    1   1.0    0    NA  0   1.9   5     1  0             3
#> 4    1   2.0    0    NA  0   3.3   5     1  0             4
#> 5    1   3.0    0    NA  0   6.6   5     1  0             5
#> 6    1   6.0    0    NA  0   9.1   5     1  0             6
#> 7    1   9.0    0    NA  0  10.8   5     1  0             7
#> 8    1  12.0    0    NA  0   8.6   5     1  0             8
#> 9    1  24.0    0    NA  0   5.6   5     1  0             9
#> 10   1  24.0    0    NA  0  44.0   6     2  0            10
#> 11   1  36.0    0    NA  0   4.0   5     1  0            11
#> 12   1  36.0    0    NA  0  27.0   6     2  0            12
#> 13   1  48.0    0    NA  0   2.7   5     1  0            13
#> 14   1  48.0    0    NA  0  28.0   6     2  0            14
#> 15   1  72.0    0    NA  0   0.8   5     1  0            15
#> 16   1  72.0    0    NA  0  31.0   6     2  0            16
#> 17   1  96.0    0    NA  0  60.0   6     2  0            17
#> 18   1 120.0    0    NA  0  65.0   6     2  0            18
#> 19   1 144.0    0    NA  0  71.0   6     2  0            19
#> 20   2   0.0    1 100.0  0    NA   1     0  0            20
#> 21   2   0.0    0    NA  0 100.0   6     2  0            21
#> 22   2  24.0    0    NA  0   9.2   5     1  0            22
#> 23   2  24.0    0    NA  0  49.0   6     2  0            23
#> 24   2  36.0    0    NA  0   8.5   5     1  0            24
#> 25   2  36.0    0    NA  0  32.0   6     2  0            25
#> 26   2  48.0    0    NA  0   6.4   5     1  0            26
#> 27   2  48.0    0    NA  0  26.0   6     2  0            27
#> 28   2  72.0    0    NA  0   4.8   5     1  0            28
#> 29   2  72.0    0    NA  0  22.0   6     2  0            29
#> 30   2  96.0    0    NA  0   3.1   5     1  0            30
#> 31   2  96.0    0    NA  0  28.0   6     2  0            31
#> 32   2 120.0    0    NA  0   2.5   5     1  0            32
#> 33   2 120.0    0    NA  0  33.0   6     2  0            33
#> 34   3   0.0    1 100.0  0    NA   1     0  0            34
#> 35   3   0.0    0    NA  0 100.0   6     2  0            35
#> 36   3   0.5    0    NA  0   0.0   5     1  0            36
#> 37   3   2.0    0    NA  0   8.4   5     1  0            37
#> 38   3   3.0    0    NA  0   9.7   5     1  0            38
#> 39   3   6.0    0    NA  0   9.8   5     1  0            39
#> 40   3  12.0    0    NA  0  11.0   5     1  0            40
#> 41   3  24.0    0    NA  0   8.3   5     1  0            41
#> 42   3  24.0    0    NA  0  46.0   6     2  0            42
#> 43   3  36.0    0    NA  0   7.7   5     1  0            43
#> 44   3  36.0    0    NA  0  22.0   6     2  0            44
#> 45   3  48.0    0    NA  0   6.3   5     1  0            45
#> 46   3  48.0    0    NA  0  19.0   6     2  0            46
#> 47   3  72.0    0    NA  0   4.1   5     1  0            47
#> 48   3  72.0    0    NA  0  20.0   6     2  0            48
#> 49   3  96.0    0    NA  0   3.0   5     1  0            49
#> 50   3  96.0    0    NA  0  42.0   6     2  0            50
#> 51   3 120.0    0    NA  0   1.4   5     1  0            51
#> 52   3 120.0    0    NA  0  49.0   6     2  0            52
#> 53   3 144.0    0    NA  0  54.0   6     2  0            53
#> 54   4   0.0    1 120.0  0    NA   1     0  0            54
#> 55   4   0.0    0    NA  0 100.0   6     2  0            55
#> 56   4   3.0    0    NA  0  12.0   5     1  0            56
#> 57   4   6.0    0    NA  0  13.2   5     1  0            57
#> 58   4   9.0    0    NA  0  14.4   5     1  0            58
#> 59   4  24.0    0    NA  0   9.6   5     1  0            59
#> 60   4  24.0    0    NA  0  30.0   6     2  0            60
#> 61   4  36.0    0    NA  0   8.2   5     1  0            61
#> 62   4  36.0    0    NA  0  24.0   6     2  0            62
#> 63   4  48.0    0    NA  0   7.8   5     1  0            63
#> 64   4  48.0    0    NA  0  13.0   6     2  0            64
#> 65   4  72.0    0    NA  0   5.8   5     1  0            65
#> 66   4  72.0    0    NA  0   9.0   6     2  0            66
#> 67   4  96.0    0    NA  0   4.3   5     1  0            67
#> 68   4  96.0    0    NA  0   9.0   6     2  0            68
#> 69   4 120.0    0    NA  0   3.0   5     1  0            69
#> 70   4 120.0    0    NA  0  11.0   6     2  0            70
#> 71   4 144.0    0    NA  0  12.0   6     2  0            71
#> 72   5   0.0    1  60.0  0    NA   1     0  0            72
#> 73   5   0.0    0    NA  0  82.0   6     2  0            73
#> 74   5   3.0    0    NA  0  11.1   5     1  0            74
#> 75   5   6.0    0    NA  0  11.9   5     1  0            75
#> 76   5   9.0    0    NA  0   9.8   5     1  0            76
#> 77   5  12.0    0    NA  0  11.0   5     1  0            77
#> 78   5  24.0    0    NA  0   8.5   5     1  0            78
#> 79   5  24.0    0    NA  0  43.0   6     2  0            79
#> 80   5  36.0    0    NA  0   7.6   5     1  0            80
#> 81   5  36.0    0    NA  0  25.0   6     2  0            81
#> 82   5  48.0    0    NA  0   5.4   5     1  0            82
#> 83   5  48.0    0    NA  0  18.0   6     2  0            83
#> 84   5  72.0    0    NA  0   4.5   5     1  0            84
#> 85   5  72.0    0    NA  0  17.0   6     2  0            85
#> 86   5  96.0    0    NA  0   3.3   5     1  0            86
#> 87   5  96.0    0    NA  0  23.0   6     2  0            87
#> 88   5 120.0    0    NA  0   2.3   5     1  0            88
#> 89   5 120.0    0    NA  0  29.0   6     2  0            89
#> 90   5 144.0    0    NA  0  41.0   6     2  0            90
#> 91   6   0.0    1 113.0  0    NA   1     0  0            91
#> 92   6   0.0    0    NA  0 100.0   6     2  0            92
#> 93   6   6.0    0    NA  0   8.6   5     1  0            93
#> 94   6  12.0    0    NA  0   8.6   5     1  0            94
#> 95   6  24.0    0    NA  0   7.0   5     1  0            95
#> 96   6  24.0    0    NA  0  34.0   6     2  0            96
#> 97   6  36.0    0    NA  0   5.7   5     1  0            97
#> 98   6  36.0    0    NA  0  23.0   6     2  0            98
#> 99   6  48.0    0    NA  0   4.7   5     1  0            99
#> 100  6  48.0    0    NA  0  20.0   6     2  0           100
#> 101  6  72.0    0    NA  0   3.3   5     1  0           101
#> 102  6  72.0    0    NA  0  16.0   6     2  0           102
#> 103  6  96.0    0    NA  0   2.3   5     1  0           103
#> 104  6  96.0    0    NA  0  17.0   6     2  0           104
#> 105  6 120.0    0    NA  0   1.7   5     1  0           105
#> 106  6 120.0    0    NA  0  18.0   6     2  0           106
#> 107  6 144.0    0    NA  0  25.0   6     2  0           107
#> 108  7   0.0    1  90.0  0    NA   1     0  0           108
#> 109  7   3.0    0    NA  0  13.4   5     1  0           109
#> 110  7   6.0    0    NA  0  12.4   5     1  0           110
#> 111  7   9.0    0    NA  0  12.7   5     1  0           111
#> 112  7  12.0    0    NA  0   8.8   5     1  0           112
#> 113  7  24.0    0    NA  0   6.1   5     1  0           113
#> 114  7  24.0    0    NA  0  36.0   6     2  0           114
#> 115  7  36.0    0    NA  0   3.5   5     1  0           115
#> 116  7  36.0    0    NA  0  33.0   6     2  0           116
#> 117  7  48.0    0    NA  0   1.8   5     1  0           117
#> 118  7  48.0    0    NA  0  28.0   6     2  0           118
#> 119  7  72.0    0    NA  0   1.5   5     1  0           119
#> 120  7  72.0    0    NA  0  52.0   6     2  0           120
#> 121  7  96.0    0    NA  0   1.0   5     1  0           121
#> 122  7  96.0    0    NA  0  80.0   6     2  0           122
#> 123  7 120.0    0    NA  0  90.0   6     2  0           123
#> 124  7 144.0    0    NA  0 100.0   6     2  0           124
#> 125  8   0.0    1 135.0  0    NA   1     0  0           125
#> 126  8   0.0    0    NA  0  88.0   6     2  0           126
#> 127  8   2.0    0    NA  0  17.6   5     1  0           127
#> 128  8   3.0    0    NA  0  17.3   5     1  0           128
#> 129  8   6.0    0    NA  0  15.0   5     1  0           129
#> 130  8   9.0    0    NA  0  15.0   5     1  0           130
#> 131  8  12.0    0    NA  0  12.4   5     1  0           131
#> 132  8  24.0    0    NA  0   7.9   5     1  0           132
#> 133  8  24.0    0    NA  0  35.0   6     2  0           133
#> 134  8  36.0    0    NA  0   7.9   5     1  0           134
#> 135  8  36.0    0    NA  0  20.0   6     2  0           135
#> 136  8  48.0    0    NA  0   5.1   5     1  0           136
#> 137  8  48.0    0    NA  0  12.0   6     2  0           137
#> 138  8  72.0    0    NA  0   3.6   5     1  0           138
#> 139  8  72.0    0    NA  0  16.0   6     2  0           139
#> 140  8  96.0    0    NA  0   2.4   5     1  0           140
#> 141  8  96.0    0    NA  0  23.0   6     2  0           141
#> 142  8 120.0    0    NA  0   2.0   5     1  0           142
#> 143  8 120.0    0    NA  0  36.0   6     2  0           143
#> 144  8 144.0    0    NA  0  48.0   6     2  0           144
#> 145  9   0.0    1  75.0  0    NA   1     0  0           145
#> 146  9   0.0    0    NA  0  92.0   6     2  0           146
#> 147  9   0.5    0    NA  0   0.0   5     1  0           147
#> 148  9   1.0    0    NA  0   1.0   5     1  0           148
#> 149  9   2.0    0    NA  0   4.6   5     1  0           149
#> 150  9   3.0    0    NA  0  12.7   5     1  0           150
#> 151  9   3.0    0    NA  0   8.0   5     1  0           151
#> 152  9   6.0    0    NA  0  12.7   5     1  0           152
#> 153  9   6.0    0    NA  0  11.5   5     1  0           153
#> 154  9   9.0    0    NA  0  12.9   5     1  0           154
#> 155  9   9.0    0    NA  0  11.4   5     1  0           155
#> 156  9  12.0    0    NA  0  11.4   5     1  0           156
#> 157  9  12.0    0    NA  0  11.0   5     1  0           157
#> 158  9  24.0    0    NA  0   9.1   5     1  0           158
#> 159  9  24.0    0    NA  0  33.0   6     2  0           159
#> 160  9  36.0    0    NA  0   8.2   5     1  0           160
#> 161  9  36.0    0    NA  0  22.0   6     2  0           161
#> 162  9  48.0    0    NA  0   5.9   5     1  0           162
#> 163  9  48.0    0    NA  0  16.0   6     2  0           163
#> 164  9  72.0    0    NA  0   3.6   5     1  0           164
#> 165  9  72.0    0    NA  0  18.0   6     2  0           165
#> 166  9  96.0    0    NA  0   1.7   5     1  0           166
#> 167  9  96.0    0    NA  0  32.0   6     2  0           167
#> 168  9 120.0    0    NA  0   1.1   5     1  0           168
#> 169  9 120.0    0    NA  0  30.0   6     2  0           169
#> 170  9 144.0    0    NA  0  45.0   6     2  0           170
#> 171 10   0.0    1 105.0  0    NA   1     0  0           171
#> 172 10   0.0    0    NA  0  90.0   6     2  0           172
#> 173 10  24.0    0    NA  0   8.6   5     1  0           173
#> 174 10  24.0    0    NA  0  39.0   6     2  0           174
#> 175 10  36.0    0    NA  0   8.0   5     1  0           175
#> 176 10  36.0    0    NA  0  22.0   6     2  0           176
#> 177 10  48.0    0    NA  0   6.0   5     1  0           177
#> 178 10  48.0    0    NA  0  17.0   6     2  0           178
#> 179 10  72.0    0    NA  0   4.4   5     1  0           179
#> 180 10  72.0    0    NA  0  17.0   6     2  0           180
#> 181 10  96.0    0    NA  0   3.6   5     1  0           181
#> 182 10  96.0    0    NA  0  22.0   6     2  0           182
#> 183 10 120.0    0    NA  0   2.8   5     1  0           183
#> 184 10 120.0    0    NA  0  25.0   6     2  0           184
#> 185 10 144.0    0    NA  0  33.0   6     2  0           185
#> 186 11   0.0    1 123.0  0    NA   1     0  0           186
#> 187 11   0.0    0    NA  0 100.0   6     2  0           187
#> 188 11   1.5    0    NA  0  11.4   5     1  0           188
#> 189 11   3.0    0    NA  0  15.4   5     1  0           189
#> 190 11   6.0    0    NA  0  17.5   5     1  0           190
#> 191 11  12.0    0    NA  0  14.0   5     1  0           191
#> 192 11  24.0    0    NA  0   9.0   5     1  0           192
#> 193 11  24.0    0    NA  0  37.0   6     2  0           193
#> 194 11  36.0    0    NA  0   8.9   5     1  0           194
#> 195 11  36.0    0    NA  0  24.0   6     2  0           195
#> 196 11  48.0    0    NA  0   6.6   5     1  0           196
#> 197 11  48.0    0    NA  0  14.0   6     2  0           197
#> 198 11  72.0    0    NA  0   4.2   5     1  0           198
#> 199 11  72.0    0    NA  0  11.0   6     2  0           199
#> 200 11  96.0    0    NA  0   3.6   5     1  0           200
#> 201 11  96.0    0    NA  0  14.0   6     2  0           201
#> 202 11 120.0    0    NA  0   2.6   5     1  0           202
#> 203 11 120.0    0    NA  0  23.0   6     2  0           203
#> 204 11 144.0    0    NA  0  33.0   6     2  0           204
#> 205 12   0.0    1 113.0  0    NA   1     0  0           205
#> 206 12   0.0    0    NA  0  85.0   6     2  0           206
#> 207 12   1.5    0    NA  0   0.6   5     1  0           207
#> 208 12   3.0    0    NA  0   2.8   5     1  0           208
#> 209 12   6.0    0    NA  0  13.8   5     1  0           209
#> 210 12   9.0    0    NA  0  15.0   5     1  0           210
#> 211 12  24.0    0    NA  0  10.5   5     1  0           211
#> 212 12  24.0    0    NA  0  25.0   6     2  0           212
#> 213 12  36.0    0    NA  0   9.1   5     1  0           213
#> 214 12  36.0    0    NA  0  15.0   6     2  0           214
#> 215 12  48.0    0    NA  0   6.6   5     1  0           215
#> 216 12  48.0    0    NA  0  11.0   6     2  0           216
#> 217 12  72.0    0    NA  0   4.9   5     1  0           217
#> 218 12  96.0    0    NA  0   2.4   5     1  0           218
#> 219 12 120.0    0    NA  0   1.9   5     1  0           219
#> 220 13   0.0    1 113.0  0    NA   1     0  0           220
#> 221 13   0.0    0    NA  0  88.0   6     2  0           221
#> 222 13   1.5    0    NA  0   3.6   5     1  0           222
#> 223 13   3.0    0    NA  0  12.9   5     1  0           223
#> 224 13   6.0    0    NA  0  12.9   5     1  0           224
#> 225 13   9.0    0    NA  0  10.2   5     1  0           225
#> 226 13  24.0    0    NA  0   6.4   5     1  0           226
#> 227 13  24.0    0    NA  0  41.0   6     2  0           227
#> 228 13  36.0    0    NA  0   6.9   5     1  0           228
#> 229 13  36.0    0    NA  0  23.0   6     2  0           229
#> 230 13  48.0    0    NA  0   4.5   5     1  0           230
#> 231 13  48.0    0    NA  0  16.0   6     2  0           231
#> 232 13  72.0    0    NA  0   3.2   5     1  0           232
#> 233 13  72.0    0    NA  0  14.0   6     2  0           233
#> 234 13  96.0    0    NA  0   2.4   5     1  0           234
#> 235 13  96.0    0    NA  0  18.0   6     2  0           235
#> 236 13 120.0    0    NA  0   1.3   5     1  0           236
#> 237 13 120.0    0    NA  0  22.0   6     2  0           237
#> 238 13 144.0    0    NA  0  35.0   6     2  0           238
#> 239 14   0.0    1  75.0  0    NA   1     0  0           239
#> 240 14   0.0    0    NA  0  85.0   6     2  0           240
#> 241 14   0.5    0    NA  0   0.0   5     1  0           241
#> 242 14   1.0    0    NA  0   2.7   5     1  0           242
#> 243 14   2.0    0    NA  0  11.6   5     1  0           243
#> 244 14   3.0    0    NA  0  11.6   5     1  0           244
#> 245 14   6.0    0    NA  0  11.3   5     1  0           245
#> 246 14   9.0    0    NA  0   9.7   5     1  0           246
#> 247 14  24.0    0    NA  0   6.5   5     1  0           247
#> 248 14  24.0    0    NA  0  32.0   6     2  0           248
#> 249 14  36.0    0    NA  0   5.2   5     1  0           249
#> 250 14  36.0    0    NA  0  22.0   6     2  0           250
#> 251 14  48.0    0    NA  0   3.6   5     1  0           251
#> 252 14  48.0    0    NA  0  21.0   6     2  0           252
#> 253 14  72.0    0    NA  0   2.4   5     1  0           253
#> 254 14  72.0    0    NA  0  28.0   6     2  0           254
#> 255 14  96.0    0    NA  0   0.9   5     1  0           255
#> 256 14  96.0    0    NA  0  38.0   6     2  0           256
#> 257 14 120.0    0    NA  0  46.0   6     2  0           257
#> 258 14 144.0    0    NA  0  65.0   6     2  0           258
#> 259 15   0.0    1  85.0  0    NA   1     0  0           259
#> 260 15   0.0    0    NA  0 100.0   6     2  0           260
#> 261 15   1.0    0    NA  0   6.6   5     1  0           261
#> 262 15   3.0    0    NA  0  11.9   5     1  0           262
#> 263 15   6.0    0    NA  0  11.7   5     1  0           263
#> 264 15   9.0    0    NA  0  12.2   5     1  0           264
#> 265 15  24.0    0    NA  0   8.1   5     1  0           265
#> 266 15  24.0    0    NA  0  43.0   6     2  0           266
#> 267 15  36.0    0    NA  0   7.4   5     1  0           267
#> 268 15  36.0    0    NA  0  26.0   6     2  0           268
#> 269 15  48.0    0    NA  0   6.8   5     1  0           269
#> 270 15  48.0    0    NA  0  15.0   6     2  0           270
#> 271 15  72.0    0    NA  0   5.3   5     1  0           271
#> 272 15  72.0    0    NA  0  13.0   6     2  0           272
#> 273 15  96.0    0    NA  0   3.0   5     1  0           273
#> 274 15  96.0    0    NA  0  21.0   6     2  0           274
#> 275 15 120.0    0    NA  0   2.0   5     1  0           275
#> 276 15 120.0    0    NA  0  28.0   6     2  0           276
#> 277 15 144.0    0    NA  0  39.0   6     2  0           277
#> 278 16   0.0    1  87.0  0    NA   1     0  0           278
#> 279 16   0.0    0    NA  0 100.0   6     2  0           279
#> 280 16  24.0    0    NA  0  10.4   5     1  0           280
#> 281 16  24.0    0    NA  0  42.0   6     2  0           281
#> 282 16  36.0    0    NA  0   8.9   5     1  0           282
#> 283 16  36.0    0    NA  0  32.0   6     2  0           283
#> 284 16  48.0    0    NA  0   7.0   5     1  0           284
#> 285 16  48.0    0    NA  0  26.0   6     2  0           285
#> 286 16  72.0    0    NA  0   4.4   5     1  0           286
#> 287 16  72.0    0    NA  0  31.0   6     2  0           287
#> 288 16  96.0    0    NA  0   3.2   5     1  0           288
#> 289 16  96.0    0    NA  0  33.0   6     2  0           289
#> 290 16 120.0    0    NA  0   2.4   5     1  0           290
#> 291 16 120.0    0    NA  0  54.0   6     2  0           291
#> 292 17   0.0    1 117.0  0    NA   1     0  0           292
#> 293 17   0.0    0    NA  0 100.0   6     2  0           293
#> 294 17  24.0    0    NA  0   7.6   5     1  0           294
#> 295 17  24.0    0    NA  0  35.0   6     2  0           295
#> 296 17  36.0    0    NA  0   6.4   5     1  0           296
#> 297 17  36.0    0    NA  0  23.0   6     2  0           297
#> 298 17  48.0    0    NA  0   6.0   5     1  0           298
#> 299 17  48.0    0    NA  0  17.0   6     2  0           299
#> 300 17  72.0    0    NA  0   4.0   5     1  0           300
#> 301 17  72.0    0    NA  0  18.0   6     2  0           301
#> 302 17  96.0    0    NA  0   3.1   5     1  0           302
#> 303 17  96.0    0    NA  0  18.0   6     2  0           303
#> 304 17 120.0    0    NA  0   2.0   5     1  0           304
#> 305 17 120.0    0    NA  0  21.0   6     2  0           305
#> 306 18   0.0    1 112.0  0    NA   1     0  0           306
#> 307 18   0.0    0    NA  0 100.0   6     2  0           307
#> 308 18  24.0    0    NA  0   7.6   5     1  0           308
#> 309 18  24.0    0    NA  0  32.0   6     2  0           309
#> 310 18  36.0    0    NA  0   6.6   5     1  0           310
#> 311 18  36.0    0    NA  0  20.0   6     2  0           311
#> 312 18  48.0    0    NA  0   5.4   5     1  0           312
#> 313 18  48.0    0    NA  0  18.0   6     2  0           313
#> 314 18  72.0    0    NA  0   3.4   5     1  0           314
#> 315 18  72.0    0    NA  0  18.0   6     2  0           315
#> 316 18  96.0    0    NA  0   1.2   5     1  0           316
#> 317 18  96.0    0    NA  0  19.0   6     2  0           317
#> 318 18 120.0    0    NA  0   0.9   5     1  0           318
#> 319 18 120.0    0    NA  0  29.0   6     2  0           319
#> 320 19   0.0    1  95.5  0    NA   1     0  0           320
#> 321 19   0.0    0    NA  0 100.0   6     2  0           321
#> 322 19  24.0    0    NA  0   6.6   5     1  0           322
#> 323 19  24.0    0    NA  0  33.0   6     2  0           323
#> 324 19  36.0    0    NA  0   5.3   5     1  0           324
#> 325 19  36.0    0    NA  0  28.0   6     2  0           325
#> 326 19  48.0    0    NA  0   3.6   5     1  0           326
#> 327 19  48.0    0    NA  0  18.0   6     2  0           327
#> 328 19  72.0    0    NA  0   2.7   5     1  0           328
#> 329 19  72.0    0    NA  0  18.0   6     2  0           329
#> 330 19  96.0    0    NA  0   1.4   5     1  0           330
#> 331 19  96.0    0    NA  0  17.0   6     2  0           331
#> 332 19 120.0    0    NA  0   1.1   5     1  0           332
#> 333 19 120.0    0    NA  0  26.0   6     2  0           333
#> 334 20   0.0    1  88.5  0    NA   1     0  0           334
#> 335 20   0.0    0    NA  0 100.0   6     2  0           335
#> 336 20  24.0    0    NA  0   9.6   5     1  0           336
#> 337 20  24.0    0    NA  0  41.0   6     2  0           337
#> 338 20  36.0    0    NA  0   8.0   5     1  0           338
#> 339 20  36.0    0    NA  0  30.0   6     2  0           339
#> 340 20  48.0    0    NA  0   6.6   5     1  0           340
#> 341 20  48.0    0    NA  0  22.0   6     2  0           341
#> 342 20  72.0    0    NA  0   5.6   5     1  0           342
#> 343 20  72.0    0    NA  0  23.0   6     2  0           343
#> 344 20  96.0    0    NA  0   3.5   5     1  0           344
#> 345 20  96.0    0    NA  0  23.0   6     2  0           345
#> 346 20 120.0    0    NA  0   2.3   5     1  0           346
#> 347 20 120.0    0    NA  0  35.0   6     2  0           347
#> 348 21   0.0    1  93.0  0    NA   1     0  0           348
#> 349 21   0.0    0    NA  0 100.0   6     2  0           349
#> 350 21  24.0    0    NA  0   7.3   5     1  0           350
#> 351 21  24.0    0    NA  0  46.0   6     2  0           351
#> 352 21  36.0    0    NA  0   6.1   5     1  0           352
#> 353 21  36.0    0    NA  0  27.0   6     2  0           353
#> 354 21  48.0    0    NA  0   4.3   5     1  0           354
#> 355 21  48.0    0    NA  0  22.0   6     2  0           355
#> 356 21  72.0    0    NA  0   3.2   5     1  0           356
#> 357 21  72.0    0    NA  0  36.0   6     2  0           357
#> 358 21  96.0    0    NA  0   2.3   5     1  0           358
#> 359 21  96.0    0    NA  0  40.0   6     2  0           359
#> 360 21 120.0    0    NA  0   1.9   5     1  0           360
#> 361 21 120.0    0    NA  0  44.0   6     2  0           361
#> 362 22   0.0    1  87.0  0    NA   1     0  0           362
#> 363 22   0.0    0    NA  0 100.0   6     2  0           363
#> 364 22  24.0    0    NA  0   8.9   5     1  0           364
#> 365 22  24.0    0    NA  0  35.0   6     2  0           365
#> 366 22  36.0    0    NA  0   8.4   5     1  0           366
#> 367 22  36.0    0    NA  0  27.0   6     2  0           367
#> 368 22  48.0    0    NA  0   8.0   5     1  0           368
#> 369 22  48.0    0    NA  0  23.0   6     2  0           369
#> 370 22  72.0    0    NA  0   4.4   5     1  0           370
#> 371 22  72.0    0    NA  0  27.0   6     2  0           371
#> 372 22  96.0    0    NA  0   3.2   5     1  0           372
#> 373 22  96.0    0    NA  0  43.0   6     2  0           373
#> 374 22 120.0    0    NA  0   1.7   5     1  0           374
#> 375 22 120.0    0    NA  0  43.0   6     2  0           375
#> 376 23   0.0    1 110.0  0    NA   1     0  0           376
#> 377 23   0.0    0    NA  0 100.0   6     2  0           377
#> 378 23  24.0    0    NA  0   9.8   5     1  0           378
#> 379 23  24.0    0    NA  0  34.0   6     2  0           379
#> 380 23  36.0    0    NA  0   8.4   5     1  0           380
#> 381 23  36.0    0    NA  0  24.0   6     2  0           381
#> 382 23  48.0    0    NA  0   6.6   5     1  0           382
#> 383 23  48.0    0    NA  0  15.0   6     2  0           383
#> 384 23  72.0    0    NA  0   4.8   5     1  0           384
#> 385 23  72.0    0    NA  0  15.0   6     2  0           385
#> 386 23  96.0    0    NA  0   3.2   5     1  0           386
#> 387 23  96.0    0    NA  0  19.0   6     2  0           387
#> 388 23 120.0    0    NA  0   2.4   5     1  0           388
#> 389 23 120.0    0    NA  0  19.0   6     2  0           389
#> 390 24   0.0    1 115.0  0    NA   1     0  0           390
#> 391 24   0.0    0    NA  0  88.0   6     2  0           391
#> 392 24  24.0    0    NA  0   8.2   5     1  0           392
#> 393 24  24.0    0    NA  0  37.0   6     2  0           393
#> 394 24  36.0    0    NA  0   7.5   5     1  0           394
#> 395 24  36.0    0    NA  0  20.0   6     2  0           395
#> 396 24  48.0    0    NA  0   6.8   5     1  0           396
#> 397 24  48.0    0    NA  0  20.0   6     2  0           397
#> 398 24  72.0    0    NA  0   5.5   5     1  0           398
#> 399 24  72.0    0    NA  0  26.0   6     2  0           399
#> 400 24  96.0    0    NA  0   4.5   5     1  0           400
#> 401 24  96.0    0    NA  0  28.0   6     2  0           401
#> 402 24 120.0    0    NA  0   3.7   5     1  0           402
#> 403 24 120.0    0    NA  0  50.0   6     2  0           403
#> 404 25   0.0    1 112.0  0    NA   1     0  0           404
#> 405 25   0.0    0    NA  0 100.0   6     2  0           405
#> 406 25  24.0    0    NA  0  11.0   5     1  0           406
#> 407 25  24.0    0    NA  0  32.0   6     2  0           407
#> 408 25  36.0    0    NA  0  10.0   5     1  0           408
#> 409 25  36.0    0    NA  0  20.0   6     2  0           409
#> 410 25  48.0    0    NA  0   8.2   5     1  0           410
#> 411 25  48.0    0    NA  0  17.0   6     2  0           411
#> 412 25  72.0    0    NA  0   6.0   5     1  0           412
#> 413 25  72.0    0    NA  0  19.0   6     2  0           413
#> 414 25  96.0    0    NA  0   3.7   5     1  0           414
#> 415 25  96.0    0    NA  0  21.0   6     2  0           415
#> 416 25 120.0    0    NA  0   2.6   5     1  0           416
#> 417 25 120.0    0    NA  0  30.0   6     2  0           417
#> 418 26   0.0    1 120.0  0    NA   1     0  0           418
#> 419 26   0.0    0    NA  0 100.0   6     2  0           419
#> 420 26  24.0    0    NA  0  10.0   5     1  0           420
#> 421 26  24.0    0    NA  0  41.0   6     2  0           421
#> 422 26  36.0    0    NA  0   9.0   5     1  0           422
#> 423 26  36.0    0    NA  0  28.0   6     2  0           423
#> 424 26  48.0    0    NA  0   7.3   5     1  0           424
#> 425 26  48.0    0    NA  0  19.0   6     2  0           425
#> 426 26  72.0    0    NA  0   5.2   5     1  0           426
#> 427 26  72.0    0    NA  0  17.0   6     2  0           427
#> 428 26  96.0    0    NA  0   3.7   5     1  0           428
#> 429 26  96.0    0    NA  0  17.0   6     2  0           429
#> 430 26 120.0    0    NA  0   2.7   5     1  0           430
#> 431 26 120.0    0    NA  0  24.0   6     2  0           431
#> 432 27   0.0    1 120.0  0    NA   1     0  0           432
#> 433 27   0.0    0    NA  0 100.0   6     2  0           433
#> 434 27  24.0    0    NA  0  11.8   5     1  0           434
#> 435 27  24.0    0    NA  0  32.0   6     2  0           435
#> 436 27  36.0    0    NA  0   9.2   5     1  0           436
#> 437 27  36.0    0    NA  0  21.0   6     2  0           437
#> 438 27  48.0    0    NA  0   7.7   5     1  0           438
#> 439 27  48.0    0    NA  0  19.0   6     2  0           439
#> 440 27  72.0    0    NA  0   4.9   5     1  0           440
#> 441 27  72.0    0    NA  0  22.0   6     2  0           441
#> 442 27  96.0    0    NA  0   3.4   5     1  0           442
#> 443 27  96.0    0    NA  0  33.0   6     2  0           443
#> 444 27 120.0    0    NA  0   2.7   5     1  0           444
#> 445 27 120.0    0    NA  0  46.0   6     2  0           445
#> 446 28   0.0    1 120.0  0    NA   1     0  0           446
#> 447 28   0.0    0    NA  0 100.0   6     2  0           447
#> 448 28  24.0    0    NA  0  10.1   5     1  0           448
#> 449 28  24.0    0    NA  0  39.0   6     2  0           449
#> 450 28  36.0    0    NA  0   8.0   5     1  0           450
#> 451 28  36.0    0    NA  0  25.0   6     2  0           451
#> 452 28  48.0    0    NA  0   6.0   5     1  0           452
#> 453 28  48.0    0    NA  0  16.0   6     2  0           453
#> 454 28  72.0    0    NA  0   4.9   5     1  0           454
#> 455 28  72.0    0    NA  0  14.0   6     2  0           455
#> 456 28  96.0    0    NA  0   3.4   5     1  0           456
#> 457 28  96.0    0    NA  0  15.0   6     2  0           457
#> 458 28 120.0    0    NA  0   2.0   5     1  0           458
#> 459 28 120.0    0    NA  0  20.0   6     2  0           459
#> 460 29   0.0    1 153.0  0    NA   1     0  0           460
#> 461 29   0.0    0    NA  0  86.0   6     2  0           461
#> 462 29  24.0    0    NA  0   8.3   5     1  0           462
#> 463 29  24.0    0    NA  0  35.0   6     2  0           463
#> 464 29  36.0    0    NA  0   7.0   5     1  0           464
#> 465 29  36.0    0    NA  0  21.0   6     2  0           465
#> 466 29  48.0    0    NA  0   5.6   5     1  0           466
#> 467 29  48.0    0    NA  0  18.0   6     2  0           467
#> 468 29  72.0    0    NA  0   4.1   5     1  0           468
#> 469 29  72.0    0    NA  0  20.0   6     2  0           469
#> 470 29  96.0    0    NA  0   3.1   5     1  0           470
#> 471 29  96.0    0    NA  0  29.0   6     2  0           471
#> 472 29 120.0    0    NA  0   2.2   5     1  0           472
#> 473 29 120.0    0    NA  0  41.0   6     2  0           473
#> 474 30   0.0    1 105.0  0    NA   1     0  0           474
#> 475 30   0.0    0    NA  0 100.0   6     2  0           475
#> 476 30  24.0    0    NA  0   9.9   5     1  0           476
#> 477 30  24.0    0    NA  0  45.0   6     2  0           477
#> 478 30  36.0    0    NA  0   7.5   5     1  0           478
#> 479 30  36.0    0    NA  0  24.0   6     2  0           479
#> 480 30  48.0    0    NA  0   6.5   5     1  0           480
#> 481 30  48.0    0    NA  0  23.0   6     2  0           481
#> 482 30  72.0    0    NA  0   4.1   5     1  0           482
#> 483 30  72.0    0    NA  0  26.0   6     2  0           483
#> 484 30  96.0    0    NA  0   2.9   5     1  0           484
#> 485 30  96.0    0    NA  0  28.0   6     2  0           485
#> 486 30 120.0    0    NA  0   2.3   5     1  0           486
#> 487 30 120.0    0    NA  0  39.0   6     2  0           487
#> 488 31   0.0    1 125.0  0    NA   1     0  0           488
#> 489 31   0.0    0    NA  0 100.0   6     2  0           489
#> 490 31  24.0    0    NA  0   9.5   5     1  0           490
#> 491 31  24.0    0    NA  0  45.0   6     2  0           491
#> 492 31  36.0    0    NA  0   7.8   5     1  0           492
#> 493 31  36.0    0    NA  0  30.0   6     2  0           493
#> 494 31  48.0    0    NA  0   6.4   5     1  0           494
#> 495 31  48.0    0    NA  0  24.0   6     2  0           495
#> 496 31  72.0    0    NA  0   4.5   5     1  0           496
#> 497 31  72.0    0    NA  0  22.0   6     2  0           497
#> 498 31  96.0    0    NA  0   3.4   5     1  0           498
#> 499 31  96.0    0    NA  0  28.0   6     2  0           499
#> 500 31 120.0    0    NA  0   2.5   5     1  0           500
#> 501 31 120.0    0    NA  0  42.0   6     2  0           501
#> 502 32   0.0    1  93.0  0    NA   1     0  0           502
#> 503 32   0.0    0    NA  0 100.0   6     2  0           503
#> 504 32  24.0    0    NA  0   8.9   5     1  0           504
#> 505 32  24.0    0    NA  0  36.0   6     2  0           505
#> 506 32  36.0    0    NA  0   7.7   5     1  0           506
#> 507 32  36.0    0    NA  0  27.0   6     2  0           507
#> 508 32  48.0    0    NA  0   6.9   5     1  0           508
#> 509 32  48.0    0    NA  0  24.0   6     2  0           509
#> 510 32  72.0    0    NA  0   4.4   5     1  0           510
#> 511 32  72.0    0    NA  0  23.0   6     2  0           511
#> 512 32  96.0    0    NA  0   3.5   5     1  0           512
#> 513 32  96.0    0    NA  0  20.0   6     2  0           513
#> 514 32 120.0    0    NA  0   2.5   5     1  0           514
#> 515 32 120.0    0    NA  0  22.0   6     2  0           515